requirements.txt
numpy
pandas
matplotlib
geopandas
fiona
pyproj
rtree
shapely
descartes
GoogleNews
python-telegram-bot
python-dotenv
tweepy
python-crontab